pyproject.toml
[tool.poetry]
name = "rag-research"
version = "0.1.0"
description = ""
authors = ["shoot649854 <jojoto8845@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tiktoken = "^0.6.0"
python-dotenv = "^1.0.1"
numpy = "^1.26.4"
matplotlib = "^3.8.3"
openai = "^1.13.3"
sentence-transformers = "^2.5.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
